Adding a new `defaultbranchname` git alias
This makes a "best effort" to try and determine the name of the default
branch of a repository, by falling back to a few different schemes.

This was what I was unfortunately NOT looking forward to figuring out
when the community decided to move to "main" as the new default branch
name.

... That being said, its an important change and I support it. It was
just a bit of work to figure out. 😅

More info here:

 - https://github.com/github/renaming/
 - https://github.blog/changelog/2020-08-26-set-the-default-branch-for-newly-created-repositories/
 - https://github.blog/changelog/2020-10-01-the-default-branch-for-newly-created-repositories-is-now-main/